test_ses/pyproject.toml
[tool.poetry]
name = "test_ses"
version = "1.0.1"
description = "pythonテンプレート"
authors = ["Taku Ikegami"]

[tool.poetry.dependencies]
python = "^3.10.8"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
pylint = "^2.15.5"
pytest-cov = "^4.0.0"
toml = "^0.10.2"
black = "^22.10.0"
boto3 = "^1.26.8"
PyYAML = "^6.0"
Jinja2 = "^3.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.py$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''